Move the size test to a new Bazel target
