+++ rs/tla_instrumentation/tla_instrumentation/tests/structs.rs
@@ -150,27 +150,6 @@ impl StructCanister {
 }
 
 #[test]
-fn size_test() {
-    let myval = TlaValue::Record(BTreeMap::from([
-        (
-            "field1".to_string(),
-            TlaValue::Function(BTreeMap::from([(
-                1_u64.to_tla_value(),
-                true.to_tla_value(),
-            )])),
-        ),
-        (
-            "field2".to_string(),
-            TlaValue::Variant {
-                tag: "tag".to_string(),
-                value: Box::new("abc".to_tla_value()),
-            },
-        ),
-    ]));
-    assert_eq!(myval.size(), 6);
-}
-
-#[test]
 fn struct_test() {
     unsafe {
         let canister = &mut *addr_of_mut!(GLOBAL);
